perf(sqlite): isolate coordinator imports from native connections
