+++ scripts/check-kysely-guardrails.mts
@@ -77,7 +77,7 @@ const rawSqliteAllowPathGroups = {
     "src/transcripts/sqlite-schema.ts",
     "src/state/sqlite-schema-shape.test-support.ts",
   ],
-  "cross-process SQLite coordination locks": ["src/infra/device-identity-coordinator.ts"],
+  "cross-process SQLite coordination locks": ["src/infra/sqlite-coordinator.ts"],
   "backup snapshot maintenance": [
     "src/commands/backup-verify.ts",
     "src/infra/backup-create.ts",
