+++ src/infra/device-identity-coordinator.ts
@@ -3,10 +3,10 @@ import {
   resolveDeviceIdentityCoordinatorPath,
   resolveDeviceIdentityCoordinatorPaths,
 } from "./device-identity-coordinator-paths.js";
-import { tryAcquireExclusiveSqliteCoordinator } from "./node-sqlite.js";
 import {
   ensurePrivateSqliteCoordinatorDirectory,
   SqliteCoordinatorError,
+  tryAcquireExclusiveSqliteCoordinator,
 } from "./sqlite-coordinator.js";
 import {
   acquireStateDatabaseCoordinator,
