+++ src/infra/node-sqlite.ts
@@ -4,7 +4,6 @@ import path from "node:path";
 import { pathToFileURL } from "node:url";
 import { formatErrorMessage } from "./errors.js";
 import { isSqliteWalResetSafeVersion } from "./sqlite-runtime-version.js";
-import { isSqliteLockError } from "./sqlite-transaction.js";
 import { installProcessWarningFilter } from "./warning-filter.js";
 
 const require = createRequire(import.meta.url);
@@ -121,48 +120,4 @@ export function readSqliteDataVersion(database: import("node:sqlite").DatabaseSy
     throw new Error("SQLite did not return a numeric PRAGMA data_version");
   }
   return row.data_version;
-}
-
-/** Hold a raw exclusive transaction until release for cross-process coordination. */
-export function tryAcquireExclusiveSqliteCoordinator(
-  location: string,
-  options: { busyTimeoutMs?: number } = {},
-): { release: () => void } | null {
-  const busyTimeoutMs = Math.max(0, Math.trunc(options.busyTimeoutMs ?? 0));
-  const database = openNodeSqliteDatabase(location);
-  try {
-    // Kysely transaction callbacks cannot own a lock beyond their synchronous commit section.
-    // This handle never writes or commits data. Keep the empty database's initial
-    // journal in memory so acquiring a lock does not create filesystem artifacts.
-    database.exec(
-      `PRAGMA busy_timeout = ${busyTimeoutMs}; PRAGMA journal_mode = MEMORY; BEGIN EXCLUSIVE;`,
-    );
-  } catch (error) {
-    database.close();
-    if (isSqliteLockError(error)) {
-      return null;
-    }
-    throw error;
-  }
-  return {
-    release: () => {
-      const errors: unknown[] = [];
-      try {
-        database.exec("ROLLBACK");
-      } catch (error) {
-        errors.push(error);
-      }
-      try {
-        database.close();
-      } catch (error) {
-        errors.push(error);
-      }
-      if (errors.length === 1) {
-        throw errors[0];
-      }
-      if (errors.length > 1) {
-        throw new AggregateError(errors, "SQLite coordinator rollback and close both failed");
-      }
-    },
-  };
 }
